pyproject.toml
[tool.poetry]
name = "web-project-back"
version = "0.1.0"
description = ""
authors = ["Knova <https://github.com/KnovaUnivalle>"]
license = "MIT license"
readme = "README.md"
packages = [{include = "web_project_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
djangorestframework-simplejwt = "^5.2.2"
asgiref = "^3.6.0"
django-allauth = "^0.54.0"
firebase-admin = "^6.1.0"
google-api-python-client = "^2.88.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"
django-environ = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"